[Description] Fix QDMA TX queue page error issue.
If the TX queue page is changed before hnat_qos_shaper_ebl function,
it may cause the function to set queue config on wrong TX queue page.

In order to avoid this situation, we will check queue ID to
switch to correct TX queue page.

If without this patch, QDMA TX queue config will be wrong
when the numbers of TX queue is greater than 16.

[Release-log]
N/A

Change-Id: I42b013982885830355b4108d6bc49f2feac1ef08
Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8203881